feat: move collators in the `collators` module
